requirements.txt
numpy==2.4.1
scipy==1.16.3
tqdm==4.67.1
pillow==11.1.0
